requirements.txt
tensorflow==2.15.0
pillow
streamlit
fastapi
uvicorn[standard]